+++ Sample Code/PingAuthenticationUI/src/main/java/com/pingidentity/authenticationui/ui/SpinnerFragment.java
@@ -1,0 +1,33 @@
+package com.pingidentity.authenticationui.ui;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import com.pingidentity.authenticationui.R;
+/*
+ * See LICENSE.txt for the Ping Authentication licensing information.
+ */
+public class SpinnerFragment extends Fragment {
+
+    public SpinnerFragment() {
+        // Required empty public constructor
+    }
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             @Nullable ViewGroup container,
+                             @Nullable Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_spinner, container, false);
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+    }
+}